Replaced int with ulong for correctness

+++ 1/mul3and5sum.d
@@ -7,7 +7,7 @@ import std.traits;
 
 void main() {
   StopWatch timer;
-  int sum;
+  ulong sum;
 
   timer.start();
   foreach (i; 1 .. 1000) {
